[feature][merge] Merge of res64-eXist branch, created to change resource size handling and representation on metadata from int (32 bit) to long (64 bit). For instance, trunk previous to this commit is able to store binary resources of more than 2GB, but as the resource size is limited to an 'int', it is wrongly reported. The changes have involved the update of metadata storage format (from int to long), the translation of some method signatures and various fixes along the code. For instance, some HTTP related classes, like RESTServer or EXistServlet, have received some changes wherever it was feasible to overcome some limitations of javax.servlet.http.HttpServletRequest and javax.servlet.http.HttpServletResponse, which limit Content-Length getters and setters to 32 bit values.
As this merge changes the database signature format number, it also contains the removal of limits on permissions mask from Dmitriy.

As far as I have tested, there shouldn't be slow downs, because it does not affect query critical paths. collections.dbx slightly changes its format, and it will grow a bit because metadata piece representing the resource size has grown from 'int' to 'long'.

And now, the set of commit messages from branch:

*	First set of changes: org.exist.collections.Collection.addBinaryResource.
*	'Content-Length' header generation by hand in order to bypass limitations from Java API.
*	Binary resources size is stored now as a long value in the database metadata
	(how can I signal the format has changed?).
*	Various APIs have been overhauled, so they use now 'long' instead of 'int' for sizes.
*	XMLRPC protocol at 'describeResource' call level has been augmented, so it is at the same
	time backward compatible and it now sends the size as a Long instead as an Integer.
*	Some places where the code is creating a byte[] array now check if the size is bigger
	than Integer.MAX_VALUE. Depending on the case, it fires an Exception or it limits to
	Integer.MAX_VALUE. Anyway, those places should be revisited, because a malicious attack
	could take down the whole server.
*	VirtualTempFileInputSource has been improved, so it is able to provide a Reader
	when encoding is set.
*	VirtualTempFile is now used by RESTServer on PUT requests, so a temp file is created
	only when it is needed.
*	Database signature has been updated to reflect underlying change. Thanks to Dmitriy Shabanov!
*	contentLength workarounds.
*	Some performance fixes.
*	Some 64 bit content handling fixes.
*	Various scalability fixes.
*	[shabanovd] remove the limits on permissions mask

svn path=/trunk/eXist/; revision=13229

+++ src/org/exist/atom/modules/Query.java
@@ -201,8 +201,8 @@ public class Query extends AtomModuleBase implements Atom {
             Reader r = new InputStreamReader(request.getInputStream(),charset);
             char [] buffer = new char[4096];
             int len;
-            int count = 0;
-            int contentLength = request.getContentLength();
+            long count = 0;
+            long contentLength = request.getContentLength();
             while ((len=r.read(buffer))>=0 && count<contentLength) {
                count += len;
                builder.append(buffer,0,len);
